requirements.txt
bs4==0.0.2
google-cloud-storage==3.9.0
slack-sdk==3.40.1